requirements-tests.txt
aiohttp == 3.10.0
anyio[trio] >= 3.2.1
argon2-cffi == 23.1.0
autoflake
email_validator >=1.1.1
itsdangerous >=1.1.0,<3.0.0
mypy == 1.11.1
orjson >= 3.2.1
pytest >= 6.2.4,< 9.0.0
pytest-asyncio
pytest-cov >= 2.12.0,< 6.0.0
python-multipart >= 0.0.5
python-socketio
ruff ==0.5.5
types-dataclasses ==0.6.6
types-orjson ==3.6.2
types-redis ==4.6.0.20240425
# types
types-ujson ==5.10.0.20240515
ujson >= 4.0.1
uvicorn[standard] == 0.30.4
